refactored a large amount of code to get tests working
