+++ src/controllers/lpaDetailsController.js
@@ -1,5 +1,5 @@
 import PageController from './pageController.js'
-import { fetchLocalAuthorities } from '../services/fetchLocalAuthorities.js'
+import { fetchLocalAuthorities } from '../utils/datasetteQueries/fetchLocalAuthorities.js'
 
 class LpaDetailsController extends PageController {
   async locals (req, res, next) {
