+++ src/filters/filters.js
@@ -4,14 +4,11 @@ import validationMessageLookup from './validationMessageLookup.js'
 import toErrorList from './toErrorList.js'
 import prettifyColumnName from './prettifyColumnName.js'
 import getFullServiceName from './getFullServiceName.js'
-import { makeDatasetSlugToReadableNameFilter, createDatasetMapping } from './makeDatasetSlugToReadableNameFilter.js'
-import datasette from '../services/datasette.js'
+import { makeDatasetSlugToReadableNameFilter } from './makeDatasetSlugToReadableNameFilter.js'
 
 const { govukMarkdown } = xGovFilters
 
-const addFilters = async (nunjucksEnv) => {
-  const datasetSlugNameTable = await datasette.runQuery('select dataset, name from dataset')
-  const datasetNameMapping = createDatasetMapping(datasetSlugNameTable.rows)
+const addFilters = (nunjucksEnv, { datasetNameMapping }) => {
   const datasetSlugToReadableName = makeDatasetSlugToReadableNameFilter(datasetNameMapping)
   nunjucksEnv.addFilter('datasetSlugToReadableName', datasetSlugToReadableName)
