+++ test/unit/endpointSubmissionForm/confirmationPage.test.js
@@ -5,9 +5,8 @@ import { setupNunjucks } from '../../../src/serverSetup/nunjucks.js'
 import { runGenericPageTests } from '../generic-page.js'
 import config from '../../../config/index.js'
 import { stripWhitespace } from '../../utils/stripWhiteSpace.js'
-import { mockDataSubjects } from '../data.js'
 
-const nunjucks = setupNunjucks({ dataSubjects: mockDataSubjects })
+const nunjucks = setupNunjucks({ datasetNameMapping: new Map() })
 
 describe('Submit confirmation View', () => {
   const params = {
@@ -18,12 +17,12 @@ describe('Submit confirmation View', () => {
   const html = stripWhitespace(nunjucks.render('submit/confirmation.html', params))
 
   runGenericPageTests(html, {
-    pageTitle: 'A Mock dataset submitted - Submit planning and housing data for England',
+    pageTitle: 'mockDataset submitted - Submit planning and housing data for England',
     serviceName: config.serviceName
   })
 
   it('should render the gov uk panel', () => {
-    const regex = new RegExp('<h1 class="govuk-panel__title".*A Mock dataset submitted.*</h1>', 'g')
+    const regex = new RegExp('<h1 class="govuk-panel__title".*mockDataset submitted.*</h1>', 'g')
     expect(html).toMatch(regex)
   })
 })
