+++ test/unit/lpa-detailsPage.test.js
@@ -3,9 +3,8 @@ import { setupNunjucks } from '../../src/serverSetup/nunjucks.js'
 import { runGenericPageTests } from './generic-page.js'
 import config from '../../config/index.js'
 import { testValidationErrorMessage } from './validation-tests.js'
-import { mockDataSubjects } from './data.js'
 
-const nunjucks = setupNunjucks({ dataSubjects: mockDataSubjects })
+const nunjucks = setupNunjucks({ datasetNameMapping: new Map() })
 
 describe('Lpa-details View', () => {
   const params = {
