+++ test/unit/startPage.test.js
@@ -4,9 +4,8 @@ import { describe } from 'vitest'
 import { setupNunjucks } from '../../src/serverSetup/nunjucks.js'
 import { runGenericPageTests } from './generic-page.js'
 import config from '../../config/index.js'
-import { mockDataSubjects } from './data.js'
 
-const nunjucks = setupNunjucks({ dataSubjects: mockDataSubjects })
+const nunjucks = setupNunjucks({ datasetNameMapping: new Map() })
 
 describe('Start View', () => {
   const params = {}
